server/requirements.txt
pillow
tensorflow
fastapi[standard]
numpy
